scripts/requirements.txt
numpy>=1.19.0
opencv-python>=4.5.1
pytesseract>=0.3.8
Pillow>=8.2.0
python-dotenv>=0.19.0
scikit-image>=0.18.0
scikit-learn>=0.24.0 